Moved glue.qt.data_slice_widget to glue.viewers.common.qt.data_slice_widget
