+++ glue/app/qt/application.py
@@ -15,7 +15,7 @@ from glue import env
 from glue.main import load_plugins
 from glue.qt import get_qapp
 from glue.qt.qtutil import data_wizard, load_ui, get_icon, action
-from glue.qt.widgets.edit_subset_mode_toolbar import EditSubsetModeToolBar
+from glue.app.qt.edit_subset_mode_toolbar import EditSubsetModeToolBar
 from glue.app.qt.mdi_area import GlueMdiArea, GlueMdiSubWindow
 from glue.qt.widgets.layer_tree_widget import PlotAction, LayerTreeWidget
 from glue.qt.widgets.settings_editor import SettingsEditor
