+++ glue/app/qt/edit_subset_mode_toolbar.py
