+++ glue/plugins/dendro_viewer/qt/viewer_widget.py
@@ -6,7 +6,7 @@ from glue.external.qt import QtGui
 from glue import core
 from glue.plugins.dendro_viewer.client import DendroClient
 from glue.viewers.common.qt.toolbar import GlueToolbar
-from glue.qt.mouse_mode import PickMode
+from glue.viewers.common.qt.mouse_mode import PickMode
 from glue.qt.qtutil import load_ui
 from glue.utils.qt.widget_properties import (ButtonProperty, CurrentComboProperty,
                                        connect_bool_button, connect_current_combo)
