+++ glue/plugins/tools/spectrum_tool/qt/spectrum_tool.py
@@ -16,7 +16,7 @@ from glue.core.callback_property import add_callback, ignore_callback
 from glue.config import fit_plugin
 from glue.viewers.common.qt.toolbar import GlueToolbar
 from glue.qt.mime import LAYERS_MIME_TYPE
-from glue.qt.mouse_mode import SpectrumExtractorMode
+from glue.viewers.common.qt.mouse_mode import SpectrumExtractorMode
 from glue.qt.qtutil import load_ui
 from glue.core.qt.simpleforms import build_form_item
 from glue.utils.qt.widget_properties import CurrentComboProperty
