+++ glue/viewers/common/qt/data_slice_widget.ui
