+++ glue/viewers/common/qt/tests/test_mouse_mode.py
