+++ glue/viewers/scatter/qt/viewer_widget.py
@@ -7,7 +7,7 @@ from glue.external.qt import QtGui
 from glue import core
 from glue.viewers.scatter.client import ScatterClient
 from glue.viewers.common.qt.toolbar import GlueToolbar
-from glue.qt.mouse_mode import (RectangleMode, CircleMode,
+from glue.viewers.common.qt.mouse_mode import (RectangleMode, CircleMode,
                                 PolyMode, HRangeMode, VRangeMode)
 from glue.qt.qtutil import load_ui
 from glue.viewers.common.qt.data_viewer import DataViewer
